backend_super_market/requirements.txt
# Django and REST Framework
Django==6.0
djangorestframework==3.17.1
django-cors-headers==4.9.0
django-filter==24.3

# JWT Authentication
djangorestframework-simplejwt==5.5.1

# Database
psycopg2-binary==2.9.12

# Image handling
Pillow==12.2.0

# Environment variables
python-decouple==3.8
python-dotenv==1.2.1

# Time zone support
pytz==2026.1.post1

# For production
gunicorn==25.3.0
whitenoise==6.12.0

# WebSocket support for real-time notifications
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1

# Background tasks and scheduling
celery==5.6.3
redis==7.4.0

# Data export and reporting
openpyxl==3.1.5
xlsxwriter==3.2.9
reportlab==4.4.10

# Barcode generation
python-barcode==0.16.1

# Email
django-templated-mail==1.1.1

# API documentation
drf-spectacular==0.29.0

# Performance monitoring
django-silk==5.5.0

# Machine Learning & Advanced Analytics
scikit-learn==1.8.0
numpy==2.4.4
pandas==3.0.2
joblib==1.5.3
scipy==1.17.1

# Receipt OCR & Image Processing
pytesseract==0.3.13
opencv-python==4.13.0.92

# Currency & Localization
forex-python==1.9.2
babel==2.18.0

# Advanced Security
cryptography==46.0.7
bcrypt==5.0.0
